refactor: table components, change to full grid table, support row cache
