+++ console/packages/starwhale-ui/src/base/data-table/headers/header-cell.tsx
@@ -5,14 +5,14 @@ import { ChevronDown, ChevronUp } from 'baseui/icon'
 import { isFocusVisible } from '@/utils/focusVisible'
 import { StatefulPopover, PLACEMENT } from 'baseui/popover'
 import { StatefulMenu } from 'baseui/menu'
-import IconFont from '../../IconFont'
+import IconFont from '../../../IconFont'
 import cn from 'classnames'
-import { SortDirectionsT } from './types'
-import { SORT_DIRECTIONS } from './constants'
-import Button from '../../Button'
-import { LocaleContext } from './locales'
-import Checkbox from '../../Checkbox'
-import { themedUseStyletron } from '../../theme/styletron'
+import { SortDirectionsT } from '../types'
+import { SORT_DIRECTIONS } from '../constants'
+import Button from '../../../Button'
+import { LocaleContext } from '../locales'
+import Checkbox from '../../../Checkbox'
+import { themedUseStyletron } from '../../../theme/styletron'
 
 type HeaderCellPropsT = {
     index: number
