+++ console/packages/starwhale-ui/src/base/data-table/index.tsx
@@ -1,9 +1,8 @@
 import type * as Types from './types'
 
-export { DataTable } from './data-table'
+export { DataTable } from './data-custom-table'
 export { StatefulContainer } from './stateful-container'
 export { StatefulDataTable } from './stateful-data-table'
-export { DataTable as Unstable_DataTable } from './data-table'
 export { StatefulContainer as Unstable_StatefulContainer } from './stateful-container'
 export { StatefulDataTable as Unstable_StatefulDataTable } from './stateful-data-table'
 export { default as AnchorColumn } from './column-anchor'
